+++ js/lib/Synth/Controller.js
@@ -1,7 +1,7 @@
 var Oscillators = require('./Oscillators/Controller.js');
 var Master = require('./Master/Controller.js');
 var Adsr = require('./Adsr/Controller.js');
-var Presets = require('./Presets/Controller.js');
+var Menu = require('./Menu/Controller.js');
 var View = require('./View.js');
 
 var pluginName = "Prometheus";
@@ -15,7 +15,7 @@ module.exports = () => {
   var master = Master();
   var adsr = Adsr();
   var oscillators = Oscillators(adsr.model);
-  var presets = Presets(pluginName, adsr, master, oscillators);
+  var menu = Menu(pluginName, adsr, master, oscillators);
 
   oscillators.connect(master.inputNode);
 
@@ -23,7 +23,7 @@ module.exports = () => {
   (parentDomElement) => {
     var synthParentView = view.connectViewTo(parentDomElement);
 
-    presets.view.connectTo(synthParentView);
+    menu.view.connectTo(synthParentView);
     master.view.connectTo(synthParentView);
     adsr.view.connectTo(synthParentView);
     oscillators.view.connectTo(synthParentView);
