+++ js/lib/Synth/Menu/Controller.js
